requirements/local.txt
# Local development dependencies go here
-r base.txt

# Your custom requirements go here
# -----------------------------------------------------------------------------
django-extensions